Refactor hard-delete book
